updated commands

+++ commands/tag.js
@@ -8,6 +8,7 @@ module.exports = {
     usage: ['(tag name)'],
     cooldown: 1,
     guildOnly: true,
+    aliases: ['tags', 't'],
     execute(message, args) {
         if (args.length) {
             MongoClient.connect(mongodbase, { useUnifiedTopology: true }, async function (err, db) {
@@ -15,13 +16,37 @@ module.exports = {
                 dbInstance = db.db(currentdb);
                 const items = await dbInstance.collection("tags").findOne({ name: args.toString() });
                 if (items == null) {
-                    message.reply("there is no tag with that name.")
-                        .then(msg => {
-                            message.delete();
-                            setTimeout(() => {
-                                msg.delete();
-                            }, 5000);
-                        })
+                    const tags = await dbInstance.collection("tags").find({}).toArray();
+                    let data = "";
+                    let listCount = 0;
+                    tags.forEach(element => {
+                        if (element.name.includes(args.toString())) {
+                            data += element.name
+                            listCount++;
+                            if (listCount != tags.length) {
+                                data += ", "
+                            }
+                        }
+                    });
+                    if (data == "") {
+                        message.reply("no tags found.")
+                            .then(msg => {
+                                message.delete();
+                                setTimeout(() => {
+                                    msg.delete();
+                                }, 5000);
+                            })
+                    } else {
+                        const modEmbed = new Discord.MessageEmbed()
+                            .setColor('#8A28F7')
+                            .setTitle("Found tags:")
+                            .setDescription(data.substring(0, data.length - 2))
+                            .setTimestamp()
+                            .setFooter(package.name + ' v. ' + package.version);
+                        message.channel.send(modEmbed);
+                        db.close();
+
+                    }
                 } else {
                     message.delete();
                     message.channel.send(eval('`' + items.content + '`'))
